added instance name to intent (#2)
Intents should include the instance id associated with the tool (it
matters if the tool will send funds from wallet A or wallet B. Since by
default all InstanceAgents will bind themselves to their method tools,
`secure_tool.py` checks for the field `name` that is common to all
agents. Otherwise it checks for the string representation of the
instance. Default behavior is to use the classname and default ID
assigned by Python.

- fixed copyright on `angel_tool.py`
- secure_tool includes instance name (or default id assigned by Python)
in intent
- WorkflowManager checks instance matches in intent